Split bolt-check into cli and js API containing check/fix methods
