+++ packages/bolt-check/src/api/index.ts
@@ -1,0 +1,2 @@
+export { default as check } from "./check";
+export { default as fix } from "./fix";
