+++ packages/bolt-check/src/cli/test.ts
@@ -1,0 +1,11 @@
+describe("Cli", () => {
+  it.skip("should print errors for both external and internal dependencies at once", async () => {
+    throw new Error("write this test");
+  });
+  it.skip("should print multiple errors for multiple internal mismatches", async () => {
+    throw new Error("write this test");
+  });
+  it.skip("should print multiple errors for multiple external mismatches", async () => {
+    throw new Error("write this test");
+  });
+});
