+++ packages/bolt-check/src/constants.ts
@@ -1,0 +1,5 @@
+export const DEPENDENCY_TYPES = [
+  "dependencies",
+  "devDependencies",
+  "optionalDependencies"
+] as const;
